+++ dataprovider/foxsports/scraper/eventdata/example_test.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports"
 	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports/model"
-	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports/runner/eventdata"
-	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports/runner/matchup"
+	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports/scraper/eventdata"
+	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports/scraper/matchup"
 	eventdatautil "github.com/lightning-dabbler/sportscrape/util/runner/eventdata"
 	matchuputil "github.com/lightning-dabbler/sportscrape/util/runner/matchup"
 )
