+++ dataprovider/foxsports/scraper/eventdata/scraper.go
