+++ dataprovider/foxsports/scraper/eventdata/scraper_mlb_batting_box_score.go
