+++ dataprovider/foxsports/scraper/eventdata/scraper_mlb_pitching_box_score.go
