+++ dataprovider/foxsports/scraper/eventdata/scraper_nba_box_score_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports"
 	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports/model"
-	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports/runner/matchup"
+	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports/scraper/matchup"
 	"github.com/lightning-dabbler/sportscrape/util/runner/eventdata"
 	matchuputil "github.com/lightning-dabbler/sportscrape/util/runner/matchup"
 	"github.com/stretchr/testify/assert"
